requirements.txt
ccxt>=4.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiohttp>=3.8.5
pandas>=2.0.3
loguru>=0.7.0
